Add link to the Russia Sanctions Project
This project was started to create, and maintain a list of, plugins for
as many popular frameworks as possible, all which do the same task of
blocking Russian and Belarusian IP addresses to websites with a polite
message.

The Russia Sanctions packages all use the same html message, so that a
common message can be distributed to as many platforms as possible. They
also use a slimed down IP database and matching algorithm, making their
implementations very small and zippy.